Re-package sources to have `io.spine.examples` as a root package.
