+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/BuilderBeforeReturnRenderer.kt
@@ -1,0 +1,66 @@
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package io.spine.examples.protodata.hello
+
+import io.spine.protodata.renderer.Renderer
+import io.spine.protodata.renderer.SourceFile
+import io.spine.protodata.renderer.SourceFileSet
+import io.spine.tools.code.Java
+
+/**
+ * The [Renderer] that inserts validation methods calls into the `build`
+ * method of the message class builder.
+ */
+public class BuilderBeforeReturnRenderer(
+    /**
+     * A container with validation methods that are associated with
+     * Java source file and should be inserted into the `build` method
+     * of the message builder class.
+     */
+    private val builderValidationMethods: BuilderValidationMethods
+) : Renderer<Java>(Java.lang()) {
+
+    override fun render(sources: SourceFileSet) {
+        if (!sources.outputRoot.endsWith("java")) {
+            return
+        }
+        sources.filter {
+            builderValidationMethods.hasMethods(it.relativePath)
+        }.forEach(::insertValidationMethodsInvocation)
+    }
+
+    /**
+     * Inserts validation method calls into the `build` method
+     * of the message builder class.
+     */
+    private fun insertValidationMethodsInvocation(sourceFile: SourceFile) {
+        val builder = sourceFile.at(BuilderBeforeReturnInsertionPoint())
+            .withExtraIndentation(2)
+
+        builderValidationMethods.methods(sourceFile.relativePath)
+            .forEach(builder::add)
+    }
+}
