+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/SizeOptionView.kt
@@ -1,4 +1,5 @@
-/* * Copyright 2024, TeamDev. All rights reserved.
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,38 +23,39 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package io.spine.examples.protodata.hello
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Validation
+import io.spine.core.External
+import io.spine.core.Subscribe
+import io.spine.core.Where
+import io.spine.protobuf.AnyPacker.unpack
+import io.spine.protodata.event.FieldOptionDiscovered
+import io.spine.protodata.plugin.View
 
-dependencies {
-    // Enable field options extension.
-    api(project(":proto-extension"))
+/**
+ * Records the [ArrayOfSizeOption] options that are applied to repeated fields.
+ */
+internal class SizeOptionView : View<SizeOptionId,
+        SizeOption,
+        SizeOption.Builder>() {
 
-    // Add module with code generation plugin to ProtoData classpath.
-    protoData(project(":codegen-plugin"))
+    /**
+     * Parameters to filter the `size` option among the other options.
+     */
+    private companion object FilterParams {
+        const val FIELD_NAME = "option.name"
+        const val FIELD_VALUE = "size"
+    }
 
-    // To allow access to `ValidatingBuilder` from the generated Kotlin code.
-    implementation(Validation.runtime)
-
-    testImplementation(JUnit.runner)
-}
-
-apply {
-    plugin("io.spine.protodata")
-}
-
-protoData {
-    // Deploy the code generation plugin to ProtoData.
-    plugins(
-        "io.spine.examples.protodata.hello.ApplySizeOptionPlugin"
-    )
-}
-
-modelCompiler {
-    java {
-        codegen {
-            validation { skipValidation() }
-        }
+    @Subscribe
+    internal fun on(
+        @External @Where(
+            field = FIELD_NAME,
+            equals = FIELD_VALUE
+        )
+        event: FieldOptionDiscovered
+    ) {
+        val option = unpack(event.option.value, ArrayOfSizeOption::class.java)
+        builder().setExpression(option.value)
     }
 }
