+++ codegen-plugin/src/main/kotlin/io/spine/examples/protodata/hello/SizeOptionViewRepository.kt
@@ -1,4 +1,5 @@
-/* * Copyright 2024, TeamDev. All rights reserved.
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,38 +23,31 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+package io.spine.examples.protodata.hello
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Validation
+import io.spine.core.EventContext
+import io.spine.protodata.event.FieldOptionDiscovered
+import io.spine.protodata.plugin.ViewRepository
+import io.spine.server.route.EventRoute
+import io.spine.server.route.EventRouting
 
-dependencies {
-    // Enable field options extension.
-    api(project(":proto-extension"))
-
-    // Add module with code generation plugin to ProtoData classpath.
-    protoData(project(":codegen-plugin"))
-
-    // To allow access to `ValidatingBuilder` from the generated Kotlin code.
-    implementation(Validation.runtime)
-
-    testImplementation(JUnit.runner)
-}
-
-apply {
-    plugin("io.spine.protodata")
-}
-
-protoData {
-    // Deploy the code generation plugin to ProtoData.
-    plugins(
-        "io.spine.examples.protodata.hello.ApplySizeOptionPlugin"
-    )
-}
-
-modelCompiler {
-    java {
-        codegen {
-            validation { skipValidation() }
+/**
+ * The repository for [SizeOptionView].
+ */
+internal class SizeOptionViewRepository : ViewRepository<SizeOptionId,
+        SizeOptionView,
+        SizeOption>() {
+    override fun setupEventRouting(routing: EventRouting<SizeOptionId>) {
+        super.setupEventRouting(routing)
+        routing.route(FieldOptionDiscovered::class.java)
+        { message: FieldOptionDiscovered, _: EventContext? ->
+            EventRoute.withId(
+                sizeOptionId {
+                    filePath = message.file
+                    typeName = message.type
+                    fieldName = message.field
+                }
+            )
         }
     }
 }
