+++ codegen-plugin/src/main/proto/spine/examples/protodata/hello/size_option.proto
@@ -1,4 +1,5 @@
-/* * Copyright 2024, TeamDev. All rights reserved.
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,37 +24,37 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Validation
+syntax = "proto3";
 
-dependencies {
-    // Enable field options extension.
-    api(project(":proto-extension"))
+package spine.examples.protodata.hello;
 
-    // Add module with code generation plugin to ProtoData classpath.
-    protoData(project(":codegen-plugin"))
+import "spine/options.proto";
 
-    // To allow access to `ValidatingBuilder` from the generated Kotlin code.
-    implementation(Validation.runtime)
+option (type_url_prefix) = "type.spine.io";
+option java_package = "io.spine.examples.protodata.hello";
+option java_outer_classname = "SizeOptionProto";
+option java_multiple_files = true;
 
-    testImplementation(JUnit.runner)
+import "spine/protodata/ast.proto";
+
+// View on the `ArrayOfSizeOption`.
+message SizeOption {
+  option (entity).kind = PROJECTION;
+
+  // ID of the option.
+  SizeOptionId id = 1;
+
+  // Value of the option which is a validation expression
+  // for the size of a repeated field.
+  string expression = 2;
 }
 
-apply {
-    plugin("io.spine.protodata")
-}
+// Triplet File+Type+Field unambiguously identifies the option.
+message SizeOptionId {
 
-protoData {
-    // Deploy the code generation plugin to ProtoData.
-    plugins(
-        "io.spine.examples.protodata.hello.ApplySizeOptionPlugin"
-    )
-}
+  spine.protodata.FilePath file_path = 1;
 
-modelCompiler {
-    java {
-        codegen {
-            validation { skipValidation() }
-        }
-    }
+  spine.protodata.TypeName type_name = 2;
+
+  spine.protodata.FieldName field_name = 3;
 }
