--- integration-tests/src/test/proto/spine/protodata/hello/contact.proto
@@ -1,60 +1,0 @@
-/*
- * Copyright 2024, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-syntax = "proto3";
-
-package spine.protodata.hello;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.protodata.hello";
-option java_multiple_files = true;
-option java_outer_classname = "ContactProto";
-option java_package = "io.spine.protodata.hello";
-
-import "google/protobuf/descriptor.proto";
-import "spine/protodata/hello/options.proto";
-
-message Contact {
-  int32 element_count = 1 [(required) = true, (min).value = "1"];
-
-  repeated string phone = 2 [(required) = true, (size).value = "element_count"];
-
-  repeated Address address = 3 [(required) = true, (size).value = "element_count"];
-
-  repeated string email = 4 [(required) = true, (size).value = "element_count"];
-}
-
-message Address {
-  int32 number_of_lines = 1 [(required) = true, (min).value = "2"];
-
-  repeated string address_line = 2 [(required) = true,
-    (size).value = "10 * number_of_lines / 5 / 2 + 10 - 2 * (3 + 2)"];
-
-  string zipcode = 3 [(required) = true];
-
-  string country = 4 [(required) = true];
-}
