--- model/src/main/proto/spine/protodata/hello/tictactoe.proto
@@ -1,80 +1,0 @@
-/*
- * Copyright 2024, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-syntax = "proto3";
-
-package spine.protodata.hello;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.protodata.hello";
-option java_multiple_files = true;
-option java_outer_classname = "TicTacToeProto";
-option java_package = "io.spine.protodata.hello";
-
-import "google/protobuf/descriptor.proto";
-import "spine/protodata/hello/options.proto";
-
-// A board for TicTacToe game.
-//
-// A board may have any size but minimum 3-by-3.
-// The field `side_size` defines the size of the board's side and
-// the number of cells is checked with the `size` option so that
-// this number always equals the "side_size * side_size" value.
-//
-message Board {
-
-  // The name of the board.
-  BoardName name = 1 [(required) = true, (validate) = true];
-
-  // Board cells.
-  repeated Cell cell = 2 [(required) = true, (size).value = "side_size * side_size"];
-
-  // The size of the board.
-  //
-  // The board must have `side_size` number of cell rows,
-  // each having `side_size` number of columns,
-  // effectively making the board "square".
-  //
-  int32 side_size = 3 [(required) = true, (min).value = "3"];
-}
-
-message BoardName {
-  string value = 1 [(required) = true];
-}
-
-enum Mark {
-  M_UNDEFINED = 0;
-  CROSS = 1;
-  CIRCLE = 2;
-}
-
-message Cell {
-  oneof value {
-    bool empty = 1;
-    Mark mark = 2;
-  }
-}
