+++ proto-extension/src/main/proto/spine/examples/protodata/hello/options.proto
@@ -1,4 +1,5 @@
-/* * Copyright 2024, TeamDev. All rights reserved.
+/*
+ * Copyright 2024, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,37 +24,48 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Validation
+syntax = "proto3";
 
-dependencies {
-    // Enable field options extension.
-    api(project(":proto-extension"))
+package spine.examples.protodata.hello;
 
-    // Add module with code generation plugin to ProtoData classpath.
-    protoData(project(":codegen-plugin"))
+import "spine/options.proto";
 
-    // To allow access to `ValidatingBuilder` from the generated Kotlin code.
-    implementation(Validation.runtime)
+option (type_url_prefix) = "type.spine.io";
+option java_package = "io.spine.examples.protodata.hello";
+option java_outer_classname = "ArrayOfSizeOptionProto";
+option java_multiple_files = true;
 
-    testImplementation(JUnit.runner)
+import "google/protobuf/descriptor.proto";
+
+extend google.protobuf.FieldOptions {
+
+    // See `ArrayOfSizeOption` for details.
+    //
+    // The field index is chosen based in `spine/options.proto`,
+    // taking the next available number in Spine's reserved range.
+    //
+    ArrayOfSizeOption size = 73855;
 }
 
-apply {
-    plugin("io.spine.protodata")
-}
+// A field option applicable to `repeated` fields,
+// telling that their size should be equal
+// to some expression, involving the values
+// of other fields of the same message.
+//
+// The `value` field supports basic math operations,
+// such as `+`, `-`, `*`, `/`.
+//
+// Example:
+//
+// message Foo {
+//
+//     int32 count = 1;
+//
+//     // There must be a number of elements
+//     // twice the `count`.
+//     repeated string value = 2 [(size),value = "count * 2"];
+// }
+message ArrayOfSizeOption {
 
-protoData {
-    // Deploy the code generation plugin to ProtoData.
-    plugins(
-        "io.spine.examples.protodata.hello.ApplySizeOptionPlugin"
-    )
-}
-
-modelCompiler {
-    java {
-        codegen {
-            validation { skipValidation() }
-        }
-    }
+    string value = 1 [(required) = true];
 }
