--- proto-extension/src/main/proto/spine/protodata/hello/options.proto
@@ -1,71 +1,0 @@
-/*
- * Copyright 2024, TeamDev. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Redistribution and use in source and/or binary forms, with or without
- * modification, must retain the above copyright notice and the following
- * disclaimer.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-syntax = "proto3";
-
-package spine.protodata.hello;
-
-import "spine/options.proto";
-
-option (type_url_prefix) = "type.spine.io";
-option java_package = "io.spine.protodata.hello";
-option java_outer_classname = "ArrayOfSizeOptionProto";
-option java_multiple_files = true;
-
-import "google/protobuf/descriptor.proto";
-
-extend google.protobuf.FieldOptions {
-
-    // See `ArrayOfSizeOption` for details.
-    //
-    // The field index is chosen based in `spine/options.proto`,
-    // taking the next available number in Spine's reserved range.
-    //
-    ArrayOfSizeOption size = 73855;
-}
-
-// A field option applicable to `repeated` fields,
-// telling that their size should be equal
-// to some expression, involving the values
-// of other fields of the same message.
-//
-// The `value` field supports basic math operations,
-// such as `+`, `-`, `*`, `/`.
-//
-// Example:
-//
-// message Foo {
-//
-//     int32 count = 1;
-//
-//     // There must be a number of elements
-//     // twice the `count`.
-//     repeated string value = 2 [(size),value = "count * 2"];
-// }
-message ArrayOfSizeOption {
-
-    string value = 1 [(required) = true];
-}
